% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drCCAcombine.R
\name{drCCAcombine}
\alias{drCCAcombine}
\title{Title}
\usage{
drCCAcombine(datasets, reg = 0, nfold = 3, nrand = 50)
}
\arguments{
\item{datasets}{: list of data matrices}

\item{reg}{: regularization value}

\item{nfold}{: number of cross validations}

\item{nrand}{: number of iterations for random matrix generation}
}
\value{
proj : projection

n : data with recommended features(counts)
}
\description{
Title
}
\examples{
library(lcca)
}
