% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCCA.R
\name{gCCA}
\alias{gCCA}
\title{Generalized CCA (GCCA)}
\usage{
gCCA(datasets, reg = 0)
}
\arguments{
\item{datasets}{: list of  data matrices}

\item{reg}{: regularization parameter (0 to 1), default is 0 (no regularization, regular GCCA)}
}
\value{
eigval : canonical correlations

eigvecs : canonical variates

proj : Projected data on canonical variates

meanvec : array of columnwise means for each data matrix

white : array of whitening matrices for each data matrix
}
\description{
Generalized CCA (GCCA)
}
\examples{
library(lcca)
}
